Use combined schema for combined date field
Introduces a combined property schema for the combined date field of
work packages.

To that end, the way the schema is referenced is moved from calling the
schema on the HalResource (i.e. work package) to fetching it from the
SchemaCacheService.

As the backend only provides generic Schema resources for all resources
(e.g. TimeEntry, WorkPackage), the difference in behaviour (i.e.
determining whether a property is editable is performed within a proxy
handler. That handler is applied whenever a schema is fetched from the
Service.